refactor: re-export transport classes + rename "drivers" to "transports"
