+++ package.json
@@ -12,7 +12,7 @@
   ],
   "exports": {
     ".": "./build/index.js",
-    "./drivers/*": "./build/drivers/*.js",
+    "./transports/*": "./build/src/transports/*.js",
     "./types/*": "./build/src/types/*.js"
   },
   "scripts": {
@@ -39,6 +39,7 @@
     "@japa/runner": "^3.1.1",
     "@swc/core": "^1.4.0",
     "@testcontainers/redis": "^10.7.1",
+    "@types/node": "^20.12.5",
     "@types/object-hash": "^3.0.6",
     "c8": "^9.1.0",
     "del-cli": "^5.1.0",
