+++ tsconfig.json
@@ -4,4 +4,4 @@
     "rootDir": "./",
     "outDir": "./build"
   }
-}
+}
